updated moi nhat

+++ DoAnWeb126/Web/admin/modules/user/delete.php
@@ -11,17 +11,30 @@
         redirectAdmin("user");
     }
     /**
-     * Kiểm tra xem có sản phẩm nào bị trùng không
+     * Kiểm tra xem khách hàng đã đặt đơn hay chưa nếu chưa thì xóa thành công ngược lại thì thất bại
      */
-    $num = $db->delete("user", $id);
-    if($num > 0){
-        $_SESSION['success'] = "Xóa thành công";
+
+    $is_user = $db->fetchOne("transaction", "user_id = $id " );
+    if($is_user == NULL)
+    {
+        $num = $db->delete("user", $id);
+        if($num > 0){
+            $_SESSION['success'] = "Xóa thành công";
+            redirectAdmin("user");
+        }
+        else
+        {
+            $_SESSION['error'] = "Xóa thất bại";
+            redirectAdmin("user");
+        }
+    }
+    else
+    {
+        $_SESSION['error'] = "Chú ý người khách hàng đang có hóa đơn thanh toán không thể xóa!";
         redirectAdmin("user");
     }
-    else{
-        $_SESSION['error'] = "Xóa thất bại";
-        redirectAdmin("user");
-    }
+
+
     
 ?>
